Update README.md
It doesn't reach f_w as that is {10, x, 1, 2} it is limited by it.

Also, you know Sbiis Saibian's Hyper E (E#) notation, until I begun adding it as a notation into my game, I never realised how perfect your library is for it! Everything just fits with the hyperions and the elements in an OmegaNum's array! Easier than scientific notation to implement!